formatting structure of code
