+++ App/meter.py
@@ -1,18 +1,19 @@
+import logging
 import os
 import pika
 from support import Support
 from dotenv import load_dotenv
 load_dotenv()
 
-class PVsimulator():
 
+class Meter():
     def __init__(self):
+        self.logger = logging.getLogger("log.log")
         self.brokeHost = os.getenv('BROKER_HOST')
         self.brokePort = os.getenv('BROKER_POST')
         self.brokeCredentials = pika.PlainCredentials(os.getenv('USER_NAME'), os.getenv('PASSWORD'))
         self.brokerQueue = os.getenv('BROKER_QUEUE')
-        self.supportFunction=Support()
-
+        self.supportFunction = Support()
 
     def connection(self):
         try:
@@ -22,10 +23,10 @@ class PVsimulator():
 
             channel = connection.channel()
             channel.queue_declare(queue=self.brokerQueue)
-
-            channel.basic_consume(queue=self.brokerQueue, on_message_callback=self.receivePower, auto_ack=True)
-            channel.start_consuming()
-
+            channel.confirm_delivery()
+            self.logger.info("Connection established")
+            powerValue= self.supportFunction.powerGenerate()
+            self.publish_massage(channel,powerValue)
         except pika.exceptions.ConnectionClosedByBroker:
             self.logger.error("The connection  is terminated by Broker.")
         except pika.exceptions.ChannelError:
@@ -33,16 +34,12 @@ class PVsimulator():
         except KeyboardInterrupt as error:
             self.logger.info("User terminates the process")
 
-
-    def receivePower(self,channel,method,properties,body):
+    def publish_massage(self,channel, powerValue):
         try:
-            powerValue= float(body)
-            pvPowerValue= self.supportFunction.powerGenerate()
-            self.supportFunction.writeCSC(powerValue,pvPowerValue)
+            channel.basic_publish(exchange='',routing_key=self.brokerQueue,body=f"{powerValue}",mandatory=True,properties=pika.BasicProperties( delivery_mode = 2, ))
+            channel.close()
 
+        except pika.exceptions.UnroutableError:
+            self.logger.error("Massage can not be send.")
         except KeyboardInterrupt as error:
             self.logger.info("User terminates the process")
-
-
-
-
