+++ Dockerfile
@@ -2,7 +2,8 @@
 FROM python:3.9
 WORKDIR /code
 
-RUN apk add --no-cache gcc musl-dev linux-headers
+
+COPY requirements.txt .
 RUN pip install -r requirements.txt
 COPY . .
-CMD ["python", "-m", "main"]
+CMD ["python", "-m", "main.py"]
