+++ docker-compose.yml
@@ -6,10 +6,19 @@ services:
     ports:
       - 5674:5672
       - 15674:15672
-  meter:
-    container_name: meter
-    build: .
 
-  pvsimulator:
-    container_name: pvsimulator
+  meter:
     build: .
+    command: python -m main.py meter
+    links:
+      - "rabbitmq-server:rabbitmq-server"
+    depends_on:
+      - "rabbitmq-server"
+
+  pv_simulator:
+    build: .
+    command: python -m main.py pv_simulator
+    links:
+      - "rabbitmq-server:rabbitmq-server"
+    depends_on:
+      - "rabbitmq-server"
